+++ .circleci/config.yml
@@ -66,6 +66,9 @@ jobs:
           name: Run shared unit tests
           command: ./gradlew -PdisablePreDex --no-daemon --max-workers=2 shared:test
       - run:
+          name: Run formtest unit tests
+          command: ./gradlew -PdisablePreDex --no-daemon --max-workers=2 formstest:test
+      - run:
           name: Run async unit tests
           command: ./gradlew -PdisablePreDex --no-daemon --max-workers=2 async:testDebug
       - run:
