+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/FormDeleterTest.java
@@ -2,7 +2,7 @@ package org.odk.collect.android.formmanagement;
 
 import org.junit.Test;
 import org.odk.collect.android.itemsets.FastExternalItemsetsRepository;
-import org.odk.collect.android.support.InMemFormsRepository;
+import org.odk.collect.formstest.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.instances.Instance;
