+++ collect_app/src/test/java/org/odk/collect/android/instancemanagement/InstanceSubmitterTest.java
@@ -1,7 +1,7 @@
 package org.odk.collect.android.instancemanagement;
 
 import org.junit.Test;
-import org.odk.collect.android.support.InMemFormsRepository;
+import org.odk.collect.formstest.InMemFormsRepository;
 import org.odk.collect.forms.Form;
 import org.odk.collect.formstest.FormUtils;
