+++ collect_app/src/test/java/org/odk/collect/android/utilities/InstanceUploaderUtilsTest.java
@@ -2,7 +2,7 @@ package org.odk.collect.android.utilities;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.odk.collect.android.support.InMemFormsRepository;
+import org.odk.collect.formstest.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.instances.Instance;
