+++ formstest/src/test/java/org/odk/collect/formstest/InMemFormsRepositoryTest.java
@@ -1,9 +1,7 @@
-package org.odk.collect.android.forms;
+package org.odk.collect.formstest;
 
 import org.junit.Before;
-import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.forms.FormsRepository;
-import org.odk.collect.formstest.FormsRepositoryTest;
 import org.odk.collect.shared.TempFiles;
 
 import java.util.function.Supplier;
